Refactor options.
Also remove obsolete version of JoinNonVariantSegmentsWithVariants.
