+++ src/main/java/com/google/cloud/genomics/dataflow/pipelines/IdentityByState.java
@@ -13,13 +13,14 @@
  */
 package com.google.cloud.genomics.dataflow.pipelines;
 
-import java.io.IOException;
-import java.security.GeneralSecurityException;
 import java.util.List;
 
 import com.google.cloud.dataflow.sdk.Pipeline;
 import com.google.cloud.dataflow.sdk.io.TextIO;
+import com.google.cloud.dataflow.sdk.options.Default;
+import com.google.cloud.dataflow.sdk.options.Description;
 import com.google.cloud.dataflow.sdk.options.PipelineOptionsFactory;
+import com.google.cloud.dataflow.sdk.options.Validation;
 import com.google.cloud.dataflow.sdk.transforms.Combine;
 import com.google.cloud.dataflow.sdk.transforms.Create;
 import com.google.cloud.dataflow.sdk.transforms.ParDo;
@@ -30,12 +31,12 @@ import com.google.cloud.genomics.dataflow.functions.AlleleSimilarityCalculator;
 import com.google.cloud.genomics.dataflow.functions.CallSimilarityCalculatorFactory;
 import com.google.cloud.genomics.dataflow.functions.FormatIBSData;
 import com.google.cloud.genomics.dataflow.functions.IBSCalculator;
+import com.google.cloud.genomics.dataflow.functions.SharedMinorAllelesCalculatorFactory;
 import com.google.cloud.genomics.dataflow.functions.grpc.JoinNonVariantSegmentsWithVariants;
-import com.google.cloud.genomics.dataflow.readers.VariantReader;
 import com.google.cloud.genomics.dataflow.readers.VariantStreamer;
-import com.google.cloud.genomics.dataflow.utils.GenomicsDatasetOptions;
+import com.google.cloud.genomics.dataflow.utils.GCSOutputOptions;
 import com.google.cloud.genomics.dataflow.utils.GenomicsOptions;
-import com.google.cloud.genomics.dataflow.utils.IdentityByStateOptions;
+import com.google.cloud.genomics.dataflow.utils.ShardOptions;
 import com.google.cloud.genomics.utils.OfflineAuth;
 import com.google.cloud.genomics.utils.ShardBoundary;
 import com.google.cloud.genomics.utils.ShardUtils;
@@ -50,23 +51,48 @@ import com.google.genomics.v1.Variant;
  */
 
 public class IdentityByState {
+  
+  public static interface Options extends ShardOptions,
+    JoinNonVariantSegmentsWithVariants.Options, GCSOutputOptions {
+
+    @Description("The ID of the Google Genomics variant set this pipeline is accessing. "
+        + "Defaults to 1000 Genomes.")
+    @Default.String("10473108253681171589")
+    String getVariantSetId();
+
+    void setVariantSetId(String variantSetId);
+
+    @Description("The class that determines the strategy for calculating the similarity of alleles.")
+    @Default.Class(SharedMinorAllelesCalculatorFactory.class)
+    Class<? extends CallSimilarityCalculatorFactory> getCallSimilarityCalculatorFactory();
+
+    void setCallSimilarityCalculatorFactory(Class<? extends CallSimilarityCalculatorFactory> kls);
+
+    public static class Methods {
+      public static void validateOptions(Options options) {
+        JoinNonVariantSegmentsWithVariants.Options.Methods.validateOptions(options);
+        GCSOutputOptions.Methods.validateOptions(options);
+      }
+    }
+
+  }
+
   // TODO: https://github.com/googlegenomics/utils-java/issues/48
   private static final String VARIANT_FIELDS = "variants(start,calls(genotype,callSetName))";
-
-  public static void main(String[] args) throws IOException, GeneralSecurityException,
-      InstantiationException, IllegalAccessException {
+  
+  public static void main(String[] args) throws Exception {
     // Register the options so that they show up via --help
-    PipelineOptionsFactory.register(IdentityByStateOptions.class);
-    IdentityByStateOptions options =
-        PipelineOptionsFactory.fromArgs(args).withValidation().as(IdentityByStateOptions.class);
+    PipelineOptionsFactory.register(Options.class);
+    Options options =
+        PipelineOptionsFactory.fromArgs(args).withValidation().as(Options.class);
     // Option validation is not yet automatic, we make an explicit call here.
-    GenomicsDatasetOptions.Methods.validateOptions(options);
+    Options.Methods.validateOptions(options);
 
     OfflineAuth auth = GenomicsOptions.Methods.getGenomicsAuth(options);
     List<StreamVariantsRequest> requests = options.isAllReferences() ?
-        ShardUtils.getVariantRequests(options.getDatasetId(), ShardUtils.SexChromosomeFilter.EXCLUDE_XY,
+        ShardUtils.getVariantRequests(options.getVariantSetId(), ShardUtils.SexChromosomeFilter.EXCLUDE_XY,
             options.getBasesPerShard(), auth) :
-              ShardUtils.getVariantRequests(options.getDatasetId(), options.getReferences(), options.getBasesPerShard());
+              ShardUtils.getVariantRequests(options.getVariantSetId(), options.getReferences(), options.getBasesPerShard());
 
     Pipeline p = Pipeline.create(options);
     p.getCoderRegistry().setFallbackCoderProvider(GenericJsonCoder.PROVIDER);
@@ -96,8 +122,7 @@ public class IdentityByState {
   }
 
   private static CallSimilarityCalculatorFactory getCallSimilarityCalculatorFactory(
-      IdentityByStateOptions options) throws InstantiationException, IllegalAccessException {
+      Options options) throws InstantiationException, IllegalAccessException {
     return options.getCallSimilarityCalculatorFactory().newInstance();
   }
-
 }
