+++ src/main/java/com/google/cloud/genomics/dataflow/pipelines/ShardedBAMWriting.java
@@ -13,6 +13,19 @@
  */
 package com.google.cloud.genomics.dataflow.pipelines;
 
+import htsjdk.samtools.SAMFileHeader;
+import htsjdk.samtools.SAMRecord;
+import htsjdk.samtools.SAMRecordIterator;
+import htsjdk.samtools.SamReader;
+import htsjdk.samtools.ValidationStringency;
+
+import java.io.IOException;
+import java.security.GeneralSecurityException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.logging.Logger;
+
 import com.google.api.services.genomics.model.Read;
 import com.google.api.services.storage.Storage;
 import com.google.cloud.dataflow.sdk.Pipeline;
@@ -32,26 +45,14 @@ import com.google.cloud.genomics.dataflow.readers.bam.ReadBAMTransform;
 import com.google.cloud.genomics.dataflow.readers.bam.ReaderOptions;
 import com.google.cloud.genomics.dataflow.readers.bam.ShardingPolicy;
 import com.google.cloud.genomics.dataflow.utils.GCSOptions;
-import com.google.cloud.genomics.dataflow.utils.GenomicsDatasetOptions;
+import com.google.cloud.genomics.dataflow.utils.GCSOutputOptions;
 import com.google.cloud.genomics.dataflow.utils.GenomicsOptions;
+import com.google.cloud.genomics.dataflow.utils.ShardOptions;
 import com.google.cloud.genomics.dataflow.utils.ShardReadsTransform;
 import com.google.cloud.genomics.dataflow.writers.WriteReadsTransform;
 import com.google.cloud.genomics.utils.Contig;
 import com.google.cloud.genomics.utils.OfflineAuth;
 import com.google.common.collect.Lists;
-
-import htsjdk.samtools.SAMFileHeader;
-import htsjdk.samtools.SAMRecord;
-import htsjdk.samtools.SAMRecordIterator;
-import htsjdk.samtools.SamReader;
-import htsjdk.samtools.ValidationStringency;
-
-import java.io.IOException;
-import java.security.GeneralSecurityException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.logging.Logger;
 
 /**
  * Demonstrates loading some Reads, sharding them, writing them to various BAM files in parallel,
@@ -59,12 +60,20 @@ import java.util.logging.Logger;
  */
 public class ShardedBAMWriting {
 
-  static interface Options extends ShardReadsTransform.Options, WriteReadsTransform.Options {
+  static interface Options extends ShardOptions, ShardReadsTransform.Options,
+    WriteReadsTransform.Options, GCSOutputOptions {
     @Description("The Google Cloud Storage path to the BAM file to get reads data from")
     @Default.String("")
     String getBAMFilePath();
 
     void setBAMFilePath(String filePath);
+    
+    public static class Methods {
+      public static void validateOptions(Options options) {
+        GCSOutputOptions.Methods.validateOptions(options);
+      }
+    }
+
   }
 
   private static final Logger LOG = Logger.getLogger(ShardedBAMWriting.class.getName());
@@ -76,11 +85,13 @@ public class ShardedBAMWriting {
   private static Iterable<Contig> contigs;
 
   public static void main(String[] args) throws GeneralSecurityException, IOException {
-    // Register the options so that they show up via --help.
+    // Register the options so that they show up via --help
     PipelineOptionsFactory.register(Options.class);
-    options = PipelineOptionsFactory.fromArgs(args).withValidation().as(Options.class);
+    options = PipelineOptionsFactory.fromArgs(args)
+        .withValidation().as(Options.class);
     // Option validation is not yet automatic, we make an explicit call here.
-    GenomicsDatasetOptions.Methods.validateOptions(options);
+    Options.Methods.validateOptions(options);
+
     auth = GenomicsOptions.Methods.getGenomicsAuth(options);
     pipeline = Pipeline.create(options);
     // Register coders.
