--- src/main/java/com/google/cloud/genomics/dataflow/utils/GenomicsDatasetOptions.java
@@ -1,111 +1,0 @@
-/*
- * Copyright (C) 2014 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.genomics.dataflow.utils;
-
-import com.google.api.client.repackaged.com.google.common.base.Strings;
-import com.google.cloud.dataflow.sdk.options.Default;
-import com.google.cloud.dataflow.sdk.options.Description;
-import com.google.cloud.dataflow.sdk.util.gcsfs.GcsPath;
-import com.google.common.base.Preconditions;
-
-/**
- * A common options class for all pipelines that operate over a single dataset and write their
- * analysis to a file.
- */
-public interface GenomicsDatasetOptions extends GenomicsOptions {
-  public static class Methods {
-
-    public static void validateOptions(GenomicsDatasetOptions options) {
-      Preconditions.checkArgument(0 < options.getBinSize(), "binSize must be greater than zero");
-      if (null != options.getOutput()) {
-        try {
-          // Check that we can parse the path.
-          GcsPath valid = GcsPath.fromUri(options.getOutput());
-          // GcsPath allows for empty bucket or filename, but that doesn't make for a good output file.
-          Preconditions.checkArgument(!Strings.isNullOrEmpty(valid.getBucket()), "Bucket must be specified");
-          Preconditions.checkArgument(!Strings.isNullOrEmpty(valid.getObject()), "Filename prefix must be specified");
-        } catch (Exception x) {
-          Preconditions.checkState(false, "output must be a valid Google Cloud Storage URL (starting with gs://)");
-        }
-      }
-    }
-  }
-
-  @Description("The ID of the Google Genomics dataset this pipeline is accessing. "
-      + "Defaults to 1000 Genomes.")
-  @Default.String("10473108253681171589")
-  String getDatasetId();
-  void setDatasetId(String datasetId);
-
-  @Description("The IDs of the Google Genomics call sets this pipeline is working with, comma "
-      + "delimited.Defaults to 1000 Genomes HG00261.")
-  @Default.String("10473108253681171589-0")
-  String getCallSetIds();
-  void setCallSetIds(String callSetIds);
-
-  // TODO refactor this option so that its only present and validated for pipelines that do write output to GCS.
-  @Description("Google Cloud Storage path prefix of the files to which to write pipeline output, if applicable.")
-  String getOutput();
-  void setOutput(String output);
-
-  @Description("The IDs of the Google Genomics transcript sets this pipeline is working with, "
-      + "comma delimited. Defaults to UCSC refGene (hg19).")
-  @Default.String("CIjfoPXj9LqPlAEQ5vnql4KewYuSAQ")
-  String getTranscriptSetIds();
-  void setTranscriptSetIds(String transcriptSetIds);
-
-  @Description("The IDs of the Google Genomics variant annotation sets this pipeline is working "
-      + "with, comma delimited. Defaults to ClinVar (GRCh37).")
-  @Default.String("CILSqfjtlY6tHxC0nNH-4cu-xlQ")
-  String getVariantAnnotationSetIds();
-  void setVariantAnnotationSetIds(String variantAnnotationSetIds);
-
-  @Description("By default, variants analyses will be run on BRCA1.  Pass this flag to run on all "
-      + "references present in the dataset.  Note that certain jobs such as PCA and IBS "
-      + "will automatically exclude X and Y chromosomes when this option is true.")
-  boolean isAllReferences();
-  void setAllReferences(boolean allReferences);
-
-  @Description("Comma separated tuples of reference:start:end,... "
-  + "Defaults to '17:41196311:41277499' for BRCA1.")
-  @Default.String("17:41196311:41277499")
-  String getReferences();
-  void setReferences(String references);
-
-  @Description("The maximum number of bases per shard.")
-  @Default.Long(1000000)
-  long getBasesPerShard();
-  void setBasesPerShard(long basesPerShard);
-
-  @Deprecated // Remove this option when fully migrated to gRPC.
-  @Description("The maximum number of calls to return. However, at least one variant will always "
-      + "be returned, even if it has more calls than this limit.  If unspecified, the default "
-      + "number of calls will be returned.")
-  @Default.Integer(0)
-  int getMaxCalls();
-  void setMaxCalls(int maxCalls);
-
-  @Description("If querying a dataset with non-variant segments (such as Complete Genomics data "
-      + "or data in Genome VCF (gVCF) format), specify this flag so that the pipeline correctly "
-      + "takes into account non-variant segment records that overlap variants within the dataset.")
-  @Default.Boolean(false)
-  boolean getHasNonVariantSegments();
-  void setHasNonVariantSegments(boolean hasNonVariantSegments);
-
-  @Description("Genomic window \"bin\" size to use for data containing non-variant segments when "
-      + "joining those non-variant segment records with variant records.")
-  @Default.Integer(1000)
-  int getBinSize();
-  void setBinSize(int binSize);
-}
